flamethrower

+++ RogueLike/Assets/Scripts/Weapons/DamageSource.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class DamageSource : MonoBehaviour
 {
-    [SerializeField] private int damageAmount = 1;
+    [SerializeField] private float damageAmount = 1f;
     private void OnTriggerEnter2D(Collider2D collision)
     {
         if (collision.gameObject.GetComponent<EnemyMovement>())
